# services/orchestrator-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: MentatLab Orchestrator API
  description: |
    RESTful API for the MentatLab orchestrator service.

    ## Overview
    The orchestrator manages agent workflow execution as DAGs (Directed Acyclic Graphs).
    It provides endpoints for:
    - **Runs**: Create, start, cancel, and monitor workflow executions
    - **Agents**: Register and manage agent definitions
    - **Flows**: Store and retrieve reusable workflow definitions
    - **Artifacts**: Upload and download run artifacts
    - **Jobs**: Monitor Kubernetes job status (when running in K8s mode)

    ## Event Streaming
    Run events are streamed via Server-Sent Events (SSE) at `/api/v1/runs/{id}/events`.
    Events include logs, checkpoints, node status changes, progress updates, and control flow events.

    ## Authentication
    Currently no authentication is required. In production, requests should be
    proxied through the gateway which handles authentication.
  version: 1.0.0
  contact:
    name: MentatLab Team
  license:
    name: MIT
servers:
  - url: http://localhost:7070
    description: Local development
  - url: http://orchestrator:7070
    description: Kubernetes internal
...
paths:
...
  /api/v1/runs:
    post:
      tags: [runs]
      summary: Create a new run
      description: Creates a new run with an optional execution plan. Use `auto_start` to immediately begin execution.
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '201':
          description: Run created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [runs]
      summary: List runs
      description: Returns a paginated list of run IDs
      operationId: listRuns
      parameters:
        - name: limit
          in: query
          description: Maximum number of results (max 500)
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRunsResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/runs/{id}/start:
    post:
      tags: [runs]
      summary: Start run execution
      description: Starts executing a previously created run
      operationId: startRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Run started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartRunResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/agents:
    get:
      tags: [agents]
      summary: List registered agents
      operationId: listAgents
      parameters:
        - name: capabilities
          in: query
          description: Filter by capabilities (comma-separated)
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAgentsResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [agents]
      summary: Register a new agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Agent already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/agents/{id}:
    get:
      tags: [agents]
      summary: Get agent details
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    put:
      tags: [agents]
      summary: Update an agent
      operationId: updateAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    delete:
      tags: [agents]
      summary: Delete an agent
      operationId: deleteAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '204':
          description: Agent deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/agents/schedule:
    post:
      tags: [agents]
      summary: Schedule an agent for execution
      description: Creates and starts a run for the specified agent manifest
      operationId: scheduleAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentScheduleRequest'
      responses:
        '202':
          description: Agent scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentScheduleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/agents/validate:
    post:
      tags: [agents]
      summary: Validate an agent manifest
      operationId: validateManifest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Agent manifest to validate
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          $ref: '#/components/responses/BadRequest'

  # Flow management
...
components:
  parameters:
    RunId:
      name: id
      in: path
      required: true
      description: Run ID
      schema:
        type: string
    AgentId:
      name: id
      in: path
      required: true
      description: Agent ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Service or dependency unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable message
        details:
          type: object
          additionalProperties: true

    # Run types
...
    Plan:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeSpec'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/EdgeSpec'
    NodeSpec:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          description: Node type (agent, conditional, for_each, subflow)
        agent_id:
          type: string
        command:
          type: array
          items:
            type: string
        image:
          type: string
        env:
          type: object
          additionalProperties:
            type: string
        inputs:
          type: array
          items:
            type: string
          description: IDs of upstream nodes
        timeout:
          type: integer
          description: Timeout in nanoseconds (Go duration)
        retries:
          type: integer
        conditional:
          $ref: '#/components/schemas/ConditionalConfig'
        for_each:
          $ref: '#/components/schemas/ForEachConfig'
        subflow:
          $ref: '#/components/schemas/SubflowConfig'
    EdgeSpec:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
        to:
          type: string
    ConditionalConfig:
      type: object
      properties:
        expression:
          type: string
        then_node:
          type: string
        else_node:
          type: string
    ForEachConfig:
      type: object
      properties:
        collection:
          type: string
        item_var:
          type: string
        body_node:
          type: string
        max_parallel:
          type: integer
    SubflowConfig:
      type: object
      properties:
        flow_id:
          type: string
        inputs:
          type: object
          additionalProperties: true
    CreateRunRequest:
      type: object
      properties:
        name:
          type: string
        plan:
          $ref: '#/components/schemas/Plan'
        auto_start:
          type: boolean
          description: Start execution immediately after creation
    CreateRunResponse:
      type: object
      properties:
        runId:
          type: string
        status:
          type: string
        sse_url:
          type: string
          description: URL for SSE event stream (when auto_start is true)
    StartRunResponse:
      type: object
      properties:
        runId:
          type: string
        status:
          type: string
        sseUrl:
          type: string
    ListRunsResponse:
      type: object
      properties:
        runs:
          type: array
          items:
            type: string
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # Event types
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        image:
          type: string
        command:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAgentRequest:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        image:
          type: string
        command:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        image:
          type: string
        command:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
    ListAgentsResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        count:
          type: integer
    AgentScheduleRequest:
      type: object
      required: [agent_manifest]
      properties:
        agent_manifest:
          type: object
          additionalProperties: true
          description: Agent manifest with id, image, command, env, etc.
    AgentScheduleResponse:
      type: object
      properties:
        resource_id:
          type: string
        status:
          type: string
        stream_id:
          type: string
        sse_url:
          type: string
    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string

    # Flow types
...